Update dcm2bids-session and xnat2bids to support XNAT labels with dashes and underscores, and to support session naming
