--- dcm2bids-session/nrg_pipeline_dicomtobids/.hg/hgrc
@@ -1,14 +1,0 @@
-# example repository config (see 'hg help config' for more info)
-[paths]
-default = https://bitbucket.org/nrg_customizations/nrg_pipeline_dicomtobids
-
-# path aliases to other clones of this repo in URLs or filesystem paths
-# (see 'hg help config.paths' for more info)
-#
-# default:pushurl = ssh://jdoe@example.net/hg/jdoes-fork
-# my-fork         = ssh://jdoe@example.net/hg/jdoes-fork
-# my-clone        = /home/jdoe/jdoes-clone
-
-[ui]
-# name and email (local to this repository, optional), e.g.
-# username = Jane Doe <jdoe@example.com>
