--- dcm2bids-session/nrg_pipeline_dicomtobids/.hg/requires
@@ -1,6 +1,0 @@
-dotencode
-fncache
-generaldelta
-revlogv1
-sparserevlog
-store
