--- dcm2bids-session/nrg_pipeline_dicomtobids/.hg/undo.branch
@@ -1,1 +1,0 @@
-default
