--- dcm2bids-session/nrg_pipeline_dicomtobids/.hg/undo.desc
@@ -1,3 +1,0 @@
-0
-pull
-https://bitbucket.org/nrg_customizations/nrg_pipeline_dicomtobids
