--- dcm2bids-session/nrg_pipeline_dicomtobids/resources/catalog/DicomToBIDS/resources/dcm2nii.ini
@@ -1,44 +1,0 @@
-[BOOL]
-DebugMode=0
-UntestedFeatures=0
-UINT16toFLOAT32=1
-Verbose=0
-Anonymize=1
-AnonymizeSourceDICOM=0
-AppendAcqSeries=1
-AppendDate=1
-AppendFilename=0
-AppendPatientName=0
-AppendProtocolName=1
-AutoCrop=0
-CollapseFolders=1
-createoutputfolder=0
-CustomRename=0
-enablereorient=0
-OrthoFlipXDim=1
-EveryFile=1
-fourD=1
-Gzip=0
-ManualNIfTIConv=1
-PhilipsPrecise=0
-RecursiveUseNameAppend=0
-SingleNIIFile=1
-SPM2=0
-Stack3DImagesWithSameAcqNum=1
-Swizzle4D=1
-UseGE_0021_104F=0
-
-[INT]
-MaxReorientMatrix=5000
-MinReorientMatrix=255
-RecursiveFolderDepth=5
-OutDirMode=0
-SiemensDTIUse0019If00181020atleast=15
-SiemensDTINoAngulationCorrectionIf00181020atleast=1000
-SiemensDTIStackIf00181020atleast=15
-BeginClip=0
-LastClip=0
-usePigz=0
-
-[STR]
-OutDir=/data/CNDA/home/
