--- dcm2bids-session/nrg_pipeline_dicomtobids/scripts/scripts/dcm2niix_setup.sh.sample
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-DCM2NIIX_HOME=/path/to/dcm2niix/directory
-export PATH=$DCM2NIIX_HOME:$PATH
