+++ setup-commands/xnat2bids/Dockerfile
@@ -6,4 +6,4 @@ RUN pip install \
     rm -r ${HOME}/.cache/pip
 
 ADD xnat2bids.py /usr/local/bin
-LABEL org.nrg.commands="[{\"name\": \"xnat2bids\", \"command-line\": \"xnat2bids.py /input /output\", \"image\": \"xnat/xnat2bids-setup:1.1\", \"version\": \"1.1\", \"type\": \"docker-setup\", \"description\": \"xnat2bids setup command. Transforms an XNAT session with BIDS and NIFTI resources into BIDS format.\"}]"
+LABEL org.nrg.commands="[{\"name\": \"xnat2bids\", \"description\": \"xnat2bids setup command. Transforms an XNAT session with BIDS and NIFTI resources into BIDS format.\", \"version\": \"1.2\", \"type\": \"docker-setup\", \"image\": \"radiologicskate/xnat2bids-setup:1.2\", \"command-line\": \"xnat2bids.py /input /output\"}]"
