+++ setup-commands/xnat2bids/command.json
@@ -1,8 +1,8 @@
 {
     "name": "xnat2bids",
     "description": "xnat2bids setup command. Transforms an XNAT session with BIDS and NIFTI resources into BIDS format.",
-    "version": "1.1",
+    "version": "1.2",
     "type": "docker-setup",
-    "image": "xnat/xnat2bids-setup:1.1",
+    "image": "radiologicskate/xnat2bids-setup:1.2",
     "command-line": "xnat2bids.py /input /output"
 }
